+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Client/IInstallationClientFactory.cs
@@ -1,0 +1,8 @@
+using Octokit;
+
+namespace Kysect.Shreks.Integration.Github.Client;
+
+public interface IInstallationClientFactory
+{
+    Task<GitHubClient> GetClient(long installationId);
+}
