+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/CredentialStores/GithubAppCredentialStore.cs
@@ -1,0 +1,19 @@
+using GitHubJwt;
+using Octokit;
+
+namespace Kysect.Shreks.Integration.Github.CredentialStores;
+
+public class GithubAppCredentialStore : ICredentialStore
+{
+    private readonly GitHubJwtFactory _gitHubJwtFactory;
+
+    public GithubAppCredentialStore(GitHubJwtFactory gitHubJwtFactory)
+    {
+        _gitHubJwtFactory = gitHubJwtFactory;
+    }
+
+    public Task<Credentials> GetCredentials()
+    {
+        return Task.FromResult(new Credentials(_gitHubJwtFactory.CreateEncodedJwtToken(), AuthenticationType.Bearer));
+    }
+}
