+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Entities/IActionNotifier.cs
@@ -1,0 +1,8 @@
+using Octokit.Webhooks;
+
+namespace Kysect.Shreks.Integration.Github.Entities;
+
+public interface IActionNotifier
+{ 
+    Task ApplyInComments(WebhookEvent webhookEvent, long issueNumber, string hook);
+}
