+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Entities/IShreksMemoryCache.cs
@@ -1,0 +1,8 @@
+using Microsoft.Extensions.Caching.Memory;
+
+namespace Kysect.Shreks.Integration.Github.Entities;
+
+public interface IShreksMemoryCache : IMemoryCache
+{ 
+    Task<TItem> GetOrCreateAsync<TItem>(object key, Func<ICacheEntry, Task<TItem>> factory);
+}
