+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Helpers/IShreksConfiguration.cs
@@ -1,0 +1,6 @@
+namespace Kysect.Shreks.Integration.Github.Helpers;
+
+public interface IShreksConfiguration
+{
+    void Verify();
+}
