+++ Source/Infrastructure/Integration/Kysect.Shreks.Integration.Github/Helpers/ShreksConfiguration.cs
@@ -1,0 +1,15 @@
+namespace Kysect.Shreks.Integration.Github.Helpers;
+
+public sealed class ShreksConfiguration : IShreksConfiguration
+{
+    public CacheConfiguration CacheConfiguration { get; init; }
+    public CacheEntryConfiguration CacheEntryConfiguration { get; init; }
+    public GithubConfiguration GithubConfiguration { get; init; }
+
+    public void Verify()
+    {
+        CacheConfiguration.Verify();
+        CacheEntryConfiguration.Verify();
+        GithubConfiguration.Verify();
+    }
+}
