--- Source/Integrations/Kysect.Shreks.GithubIntegration/Client/IInstallationClientFactory.cs
@@ -1,8 +1,0 @@
-using Octokit;
-
-namespace Kysect.Shreks.GithubIntegration.Client;
-
-public interface IInstallationClientFactory
-{
-    Task<GitHubClient> GetClient(long installationId);
-}
