--- Source/Integrations/Kysect.Shreks.GithubIntegration/Entities/IActionNotifier.cs
@@ -1,8 +1,0 @@
-using Octokit.Webhooks;
-
-namespace Kysect.Shreks.GithubIntegration.Entities;
-
-public interface IActionNotifier
-{ 
-    Task ApplyInComments(WebhookEvent webhookEvent, long issueNumber, string hook);
-}
