--- Source/Integrations/Kysect.Shreks.GithubIntegration/Entities/IShreksMemoryCache.cs
@@ -1,8 +1,0 @@
-using Microsoft.Extensions.Caching.Memory;
-
-namespace Kysect.Shreks.GithubIntegration.Entities;
-
-public interface IShreksMemoryCache : IMemoryCache
-{ 
-    Task<TItem> GetOrCreateAsync<TItem>(object key, Func<ICacheEntry, Task<TItem>> factory);
-}
