--- Source/Integrations/Kysect.Shreks.GithubIntegration/Helpers/IShreksConfiguration.cs
@@ -1,6 +1,0 @@
-namespace Kysect.Shreks.GithubIntegration.Helpers;
-
-public interface IShreksConfiguration
-{
-    void Verify();
-}
