+++ Source/Kysect.Shreks.sln
@@ -5,19 +5,17 @@ VisualStudioVersion = 17.3.32611.2
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Domain", "Domain", "{6BC0B5E6-94B2-4984-B201-7ACD167E783A}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Abstractions", "Domain\Kysect.Shreks.Abstractions\Kysect.Shreks.Abstractions.csproj", "{62A93622-EF0E-4353-AAA9-A0034ECD1D81}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integration", "Integration", "{D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Integrations", "Integrations", "{D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentations", "Presentations", "{7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentation", "Presentation", "{7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{D6F2860A-2B15-465E-B1BB-B8B3AE829A9C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Playground", "Tests\Kysect.Shreks.Playground\Kysect.Shreks.Playground.csproj", "{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Playground", "Playground\Kysect.Shreks.Playground\Kysect.Shreks.Playground.csproj", "{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.GithubIntegration", "Integrations\Kysect.Shreks.GithubIntegration\Kysect.Shreks.GithubIntegration.csproj", "{2D8C7DC4-E72C-4F6C-B361-576576E44EFB}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Integration.Github", "Infrastructure\Integration\Kysect.Shreks.Integration.Github\Kysect.Shreks.Integration.Github.csproj", "{2D8C7DC4-E72C-4F6C-B361-576576E44EFB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.GoogleIntegration", "Integrations\Kysect.Shreks.GoogleIntegration\Kysect.Shreks.GoogleIntegration.csproj", "{B29C5E6D-2FE4-499E-9AE6-8712CAC38CD8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Integration.Google", "Infrastructure\Integration\Kysect.Shreks.Integration.Google\Kysect.Shreks.Integration.Google.csproj", "{B29C5E6D-2FE4-499E-9AE6-8712CAC38CD8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Core", "Domain\Kysect.Shreks.Core\Kysect.Shreks.Core.csproj", "{D23139A5-29A9-449E-8AD3-2FC8BC5E8086}"
 EndProject
@@ -33,7 +31,9 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Tests", "Test
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Seeding", "Infrastructure\Kysect.Shreks.Seeding\Kysect.Shreks.Seeding.csproj", "{5DCD0DCE-A3B4-425A-A691-F001DA77D209}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.GithubIntegration.Host", "Presentations\Kysect.Shreks.GithubIntegration.Host\Kysect.Shreks.GithubIntegration.Host.csproj", "{9CE8224C-7446-4FB0-AA53-28FE4A668848}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Playground.Github", "Playground\Kysect.Shreks.Playground.Github\Kysect.Shreks.Playground.Github.csproj", "{9CE8224C-7446-4FB0-AA53-28FE4A668848}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Playground", "Playground", "{DC866D57-D41E-455D-BD7D-85F1DD575823}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -41,10 +41,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{62A93622-EF0E-4353-AAA9-A0034ECD1D81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{62A93622-EF0E-4353-AAA9-A0034ECD1D81}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{62A93622-EF0E-4353-AAA9-A0034ECD1D81}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{62A93622-EF0E-4353-AAA9-A0034ECD1D81}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -86,16 +82,16 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{62A93622-EF0E-4353-AAA9-A0034ECD1D81} = {6BC0B5E6-94B2-4984-B201-7ACD167E783A}
-		{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE} = {D6F2860A-2B15-465E-B1BB-B8B3AE829A9C}
 		{2D8C7DC4-E72C-4F6C-B361-576576E44EFB} = {D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E}
 		{B29C5E6D-2FE4-499E-9AE6-8712CAC38CD8} = {D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E}
 		{D23139A5-29A9-449E-8AD3-2FC8BC5E8086} = {6BC0B5E6-94B2-4984-B201-7ACD167E783A}
 		{2277FA49-1C47-4798-9CD9-1BB57FE6E383} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 		{DDF7FF3A-29B3-4AF1-831A-869AB641B9EE} = {C578EF62-E40C-42E7-AB9D-E7BA215F342C}
 		{6B4F926D-8DE4-4204-8B18-5EC5D42D5576} = {D6F2860A-2B15-465E-B1BB-B8B3AE829A9C}
-		{9CE8224C-7446-4FB0-AA53-28FE4A668848} = {7A93BF5E-D7A6-4D8A-AC22-9D5AF09C4880}
 		{5DCD0DCE-A3B4-425A-A691-F001DA77D209} = {C578EF62-E40C-42E7-AB9D-E7BA215F342C}
+		{DE47A0F5-9E5D-4086-88AB-33850DA4EFAE} = {DC866D57-D41E-455D-BD7D-85F1DD575823}
+		{9CE8224C-7446-4FB0-AA53-28FE4A668848} = {DC866D57-D41E-455D-BD7D-85F1DD575823}
+		{D66AA3ED-1CC1-405E-BFD3-7DBA6EDC0E2E} = {C578EF62-E40C-42E7-AB9D-E7BA215F342C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C67DB-70F8-4C47-A3F4-7C94BECE1657}
