+++ Source/Playground/Kysect.Shreks.Playground.Github/appsettings.json
