+++ Source/Playground/Kysect.Shreks.Playground/Program.cs
