--- Source/Presentations/Kysect.Shreks.GithubIntegration.Host/Kysect.Shreks.GithubIntegration.Host.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.3.2" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <ProjectReference Include="..\..\Integrations\Kysect.Shreks.GithubIntegration\Kysect.Shreks.GithubIntegration.csproj" />
-    </ItemGroup>
-
-</Project>
